restructured files, imports for elements using index were kind of weird and causing bugs for some elements, might need to look into that
