+++ src/app/components/containers/sidebar/ProgressWell.js
@@ -1,8 +1,8 @@
-import { Badge, Panel, ProgressBar } from 'react-bootstrap';
-import { Colors, Decision } from '../../../../Elements/constants';
 import React, { Component } from 'react';
-
 import { connect } from 'react-redux';
+import { Badge, Panel, ProgressBar } from 'react-bootstrap';
+
+import { Colors, Decision } from '../../../globals/constants';
 
 class ProgressWell extends Component {
     constructor(props, context) {
