+++ src/app/components/elements/index.js
@@ -1,4 +1,6 @@
 export * from './DropdownElement'
+export * from './DecisionButtonGroup'
 export * from './FormElement'
 export * from './Keywords'
+export * from './RevNavbar'
 export * from './SearchGroup'
