+++ src/app/components/presentationals/paperScreening/PaperBodyPresentational.js
@@ -1,8 +1,8 @@
-import { Col, Row } from 'react-bootstrap';
-
-import DecisionButtonGroup from '../../Elements/DecisionButtonGroup';
-import Highlighter from "react-highlight-words";
 import React from 'react';
+import { Col, Row } from 'react-bootstrap';
+import Highlighter from "react-highlight-words";
+
+import DecisionButtonGroup from '../../elements/DecisionButtonGroup';
 
 const PaperBodyPresentational = (props) => {
     return (
