+++ src/app/components/presentationals/sidebar/Sidebar.js
@@ -1,9 +1,9 @@
 import React, { Component } from 'react';
 
-import Keywords from './Containers/KeywordHighlighter';
-import LibraryCounts from './Containers/LibraryCounts';
-import ProgressWell from './Containers/ProgressWell';
-import SearchgroupsFilter from './Containers/SearchgroupsFilter';
+import Keywords from '../../containers/sidebar/KeywordHighlighter';
+import LibraryCounts from '../../containers/sidebar/LibraryCounts';
+import ProgressWell from '../../containers/sidebar/ProgressWell';
+import SearchgroupsFilter from '../../containers/sidebar/SearchgroupsFilter';
 
 export default class Sidebar extends Component {
     render() {
